notebooks/requirements.txt
streamlit==1.18.0
Pillow==9.5.0
torch==2.0.1
numpy==1.25.0
requests==2.31.0
ultralytics==8.0.26
opencv-python==4.7.0
gdown==4.7.1